파일 정리 및 공통 레이아웃 분리
