+++ weekly-mission/src/Hooks/useFetchData.jsx
@@ -2,26 +2,29 @@ import { useState, useEffect } from "react";
 
 const useFetchData = (url) => {
   const [data, setData] = useState(null);
+  const [isLoading, setIsLoading] = useState(true);
 
   useEffect(() => {
+    const fetchData = async () => {
+      setIsLoading(true);
+      try {
+        const response = await fetch(url);
+        if (!response.ok) {
+          throw new Error("Failed to fetch data");
+        }
+        const data = await response.json();
+        setData(data);
+      } catch (error) {
+        console.error(error);
+      } finally {
+        setIsLoading(false);
+      }
+    };
+
     fetchData();
   }, [url]);
 
-  const fetchData = async () => {
-    try {
-      const response = await fetch(url);
-
-      if (!response.ok) {
-        throw new Error("Failed to fetch folder data");
-      }
-      const data = await response.json();
-      setData(data);
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  return data;
+  return { data, isLoading };
 };
 
 export default useFetchData;
