+++ weekly-mission/src/components/ActionButton/ActionButton.styled.js
@@ -1,0 +1,19 @@
+import styled from "styled-components";
+
+export const ButtonWrapper = styled.div`
+  display: flex;
+  gap: 12px;
+`;
+
+export const Button = styled.button`
+  font-size: 14px;
+  font-weight: 600;
+  color: #9fa6b2;
+  background-color: white;
+  border: none;
+  background-repeat: no-repeat;
+  background-position: left center;
+  background-size: 18px;
+  padding-left: 20px;
+  cursor: pointer;
+`;
