+++ weekly-mission/src/components/AddFolderLinkBar/AddFolderLink.styled.js
@@ -1,8 +1,7 @@
 import styled from "styled-components";
-import linkIcon from "../../../../image/link.png";
-import Button from "../../shared/Button";
+import Button from "../Button/Button";
 
-const AddFolderLink_Container = styled.div`
+export const AddFolderLink_Container = styled.div`
   position: relative;
   display: flex;
   align-items: center;
@@ -10,7 +9,7 @@ const AddFolderLink_Container = styled.div`
   max-width: 800px;
 `;
 
-const InputWrapper = styled.div`
+export const InputWrapper = styled.div`
   position: relative;
   display: flex;
   align-items: center;
@@ -18,7 +17,7 @@ const InputWrapper = styled.div`
   padding: 0 20px;
 `;
 
-const Input = styled.input`
+export const Input = styled.input`
   width: 100%;
   height: 69px;
   border: 1px solid #6d6afe;
@@ -30,7 +29,7 @@ const Input = styled.input`
   outline: none;
 `;
 
-const Image = styled.img`
+export const Image = styled.img`
   position: absolute;
   top: 50%;
   left: 40px;
@@ -38,7 +37,7 @@ const Image = styled.img`
   margin-right: 10px;
 `;
 
-const AddButton = styled(Button)`
+export const AddButton = styled(Button)`
   position: absolute;
   right: 50px;
   top: 50%;
@@ -48,17 +47,3 @@ const AddButton = styled(Button)`
   font-size: 14px;
   font-weight: 600;
 `;
-
-const AddFolderLink = () => {
-  return (
-    <AddFolderLink_Container>
-      <InputWrapper>
-        <Image src={linkIcon} alt="Link Icon" />
-        <Input type="text" placeholder="링크를 추가해 보세요." />
-      </InputWrapper>
-      <AddButton>추가하기</AddButton>
-    </AddFolderLink_Container>
-  );
-};
-
-export default AddFolderLink;
