+++ weekly-mission/src/components/FolderContent/LinksContent.jsx
@@ -1,0 +1,33 @@
+import * as S from "./LinksContent.styled";
+import Card from "../Card/Card";
+import NoLink from "../NoLink/NoLinks";
+import ActionButton from "../ActionButton/ActionButton";
+
+const LinksContent = ({
+  linksData,
+  activeFolderName,
+  activeFolderId,
+  onClick,
+  shareIcon,
+}) => {
+  const isEmpty = !linksData || linksData.data.length === 0;
+
+  return (
+    <>
+      <S.FoldermenuToolbar>
+        <S.SelectedFolder>{activeFolderName}</S.SelectedFolder>
+        {activeFolderId !== null && <ActionButton onClick={onClick} />}
+      </S.FoldermenuToolbar>
+
+      <S.FolderContainer>
+        {isEmpty ? (
+          <NoLink />
+        ) : (
+          linksData.data.map((link) => <Card key={link.id} link={link} />)
+        )}
+      </S.FolderContainer>
+    </>
+  );
+};
+
+export default LinksContent;
