+++ weekly-mission/src/components/NoLink/NoLinks.jsx
@@ -1,0 +1,7 @@
+import * as S from "./Nolink.styled";
+
+const NoLink = () => {
+  return <S.NoLinkLayout>저장된 링크가 없습니다.</S.NoLinkLayout>;
+};
+
+export default NoLink;
