+++ weekly-mission/src/components/NoLink/Nolink.styled.js
@@ -1,10 +1,6 @@
 import styled from "styled-components";
 
-function NoLink() {
-  return <NoLinkLayout>저장된 링크가 없습니다.</NoLinkLayout>;
-}
-
-const NoLinkLayout = styled.div`
+export const NoLinkLayout = styled.div`
   display: flex;
   width: 1060px;
   height: 100px;
@@ -12,5 +8,3 @@ const NoLinkLayout = styled.div`
   justify-content: center;
   align-items: center;
 `;
-
-export default NoLink;
